RENAMED APP.JS AND MOVED API ROUTES TO TOURVIEW
