+++ client/src/App.css
@@ -1,7 +1,10 @@
 .App {
-	text-align: center;
+  text-align: center;
+}
+
+.TourView {
 }
 
 img {
-	width: 50%;
+  width: 50%;
 }
